Implement q_remove_head and q_remove_tail
q_remove_head and q_remove_tail remove an element from the queue.
Since they only differ in the position of the removed node,
I introduced q_remove to simplify their implementation.

The new node parameter in q_remove represents the node to be removed,
which is either "head->next" or "head->prev", depending on the caller.
This reduces code duplication and improves clarity.

Change-Id: Ic8b12040d6c184d240e867c5781d7629aef368c4